Fiddling with the hookup and handoff exchange between Ingredients and Page Three.
For some reason, the recipe cards stop rendering on Page Three.

BUT!!! The ingredients are successfully pulled off of the Ingredients page and inserted into the Recipe API.
Lines 15 - 17 are the magic.

There's also a bunch of attempted code around Lines 108 - 1125 of ATTEMPTING to create functionality for the Recipe Title to be stored to Local Storage.

It does not work. I have no idea why.

Also, the Recipe Link now has its Name automatically set to the title of the recipe. If we can just TAKE that, and store it, but ONLY STORE IT when the user clicks on the recipe they want...